+++ Clients/src/presentation/pages/Reporting/Reports/index.tsx
@@ -2,9 +2,8 @@ import React, { useState, lazy, Suspense, useEffect, useCallback } from "react";
 import { Stack, Box, Typography } from "@mui/material";
 const ReportTable = lazy(() => import("../../../components/Table/ReportTable"));
 import { TITLE_OF_COLUMNS } from "./constants";
-import useGeneratedReports, {
-  GeneratedReports,
-} from "../../../../application/hooks/useGeneratedReports";
+import useGeneratedReports from "../../../../application/hooks/useGeneratedReports";
+import { GeneratedReports } from "../../../../domain/interfaces/iReports";
 import { styles, reportTablePlaceholder } from "./styles";
 import { deleteEntityById } from "../../../../application/repository/entity.repository";
 import { handleAlert } from "../../../../application/tools/alertUtils";
